Add debuff selection and minimum damage for vibrations configs. And some fixes.
Update error handling to catch more error types.
Add some extra logging in the debug setting to try and figure out the cause of toys sometimes not stopping to vibrate.
Remove unused parameter in bVibration.Died
Fix NullReferenceError in tPlayer.PostUpdate from tSystem.tSys never being set. As a result, chat is spammed when not connected to intiface (because it tries to connect every tick now). The if-statement has temporarily been disabled.